Move builders to separate files. Renaming
