+++ mobile/src/reducers/calendar/calendar-weeks-builder.spec.ts
@@ -1,5 +1,6 @@
-import { CalendarWeeksBuilder, WeekModel } from './calendar.model';
+import { WeekModel } from './calendar.model';
 import moment from 'moment';
+import { CalendarWeeksBuilder } from './calendar-weeks-builder';
 
 describe('calendarWeeksBuilder', () => {
     describe('fillWeekWithPrevMonthDays', () => {
