+++ mobile/src/reducers/calendar/calendar-weeks-builder.ts
@@ -1,0 +1,72 @@
+import { WeekModel } from './calendar.model';
+import moment, { Moment } from 'moment';
+
+export class CalendarWeeksBuilder {
+    private readonly weeksPerPage = 6;
+    private readonly daysPerWeek = 7;
+
+    public fillWeekWithPrevMonthDays(
+        weekModel: WeekModel,
+        currentWeek: number,
+        date: Moment,
+        today: Moment
+    ) {
+        const before = moment(date);
+        before.add(-1, 'days');
+
+        while (before.week() === currentWeek) {
+
+            const day = {
+                date: moment(before),
+                today: before.isSame(today, 'day'),
+                belongsToCurrentMonth: false
+            };
+
+            weekModel.days.unshift(day);
+
+            before.add(-1, 'days');
+        }
+    }
+
+    public buildWeeks(currentMonth: number, currentYear: number): WeekModel[] {
+        const date = moment({
+            date: 1, // start filling from the first day of the month
+            month: currentMonth,
+            year: currentYear });
+
+        const currentWeek = date.week();
+
+        const weeksResult: WeekModel[] = [];
+
+        let weekIndex = 1;
+        let weekModel: WeekModel = { days: [], weekIndex };
+
+        const today = moment();
+
+        this.fillWeekWithPrevMonthDays(
+            weekModel,
+            currentWeek,
+            date,
+            today);
+
+        while (weekIndex <= this.weeksPerPage) {
+            if (weekModel.days.length === this.daysPerWeek) {
+                weeksResult.push(weekModel);
+                ++weekIndex;
+                weekModel = { days: [], weekIndex };
+            }
+
+            const day = {
+                date: moment(date),
+                today: date.isSame(today, 'day'),
+                belongsToCurrentMonth: date.month() === currentMonth
+            };
+
+            weekModel.days.push(day);
+
+            date.add(1, 'days');
+        }
+
+        return weeksResult;
+    }
+}
