+++ mobile/src/reducers/calendar/calendar.model.ts
@@ -12,77 +12,7 @@ export interface WeekModel {
     weekIndex: number;
 }
 
-export class CalendarWeeksBuilder {
-    private readonly weeksPerPage = 6;
-    private readonly daysPerWeek = 7;
-
-    public fillWeekWithPrevMonthDays(
-        weekModel: WeekModel,
-        currentWeek: number,
-        date: Moment,
-        today: Moment
-    ) {
-        const before = moment(date);
-        before.add(-1, 'days');
-
-        while (before.week() === currentWeek) {
-
-            const day = {
-                date: moment(before),
-                today: before.isSame(today, 'day'),
-                belongsToCurrentMonth: false
-            };
-
-            weekModel.days.unshift(day);
-
-            before.add(-1, 'days');
-        }
-    }
-
-    public buildWeeks(currentMonth: number, currentYear: number): WeekModel[] {
-        const date = moment({
-            date: 1, // start filling from the first day of the month
-            month: currentMonth,
-            year: currentYear });
-
-        const currentWeek = date.week();
-
-        const weeksResult: WeekModel[] = [];
-
-        let weekIndex = 1;
-        let weekModel: WeekModel = { days: [], weekIndex };
-
-        const today = moment();
-
-        this.fillWeekWithPrevMonthDays(
-            weekModel,
-            currentWeek,
-            date,
-            today);
-
-        while (weekIndex <= this.weeksPerPage) {
-            if (weekModel.days.length === this.daysPerWeek) {
-                weeksResult.push(weekModel);
-                ++weekIndex;
-                weekModel = { days: [], weekIndex };
-            }
-
-            const day = {
-                date: moment(date),
-                today: date.isSame(today, 'day'),
-                belongsToCurrentMonth: date.month() === currentMonth
-            };
-
-            weekModel.days.push(day);
-
-            date.add(1, 'days');
-        }
-
-        return weeksResult;
-    }
-}
-
-type IntervalType = 'startInterval' | 'interval' | 'endInterval' | 'intervalBoundary';
+export type IntervalType = 'startInterval' | 'interval' | 'endInterval' | 'intervalBoundary';
 
 export interface IntervalModel {
     intervalType: IntervalType;
@@ -113,43 +43,5 @@ export class IntervalsModel {
 
     public static generateKey(date: Moment): string {
         return date.format('DD-MM-YYYY');
-    }
-}
-
-export class CalendarIntervalsBuilder {
-
-    public buildIntervals(calendarEvents: CalendarEvents[]) {
-        const intervalsModel = new IntervalsModel();
-
-        for (let calendarEvent of calendarEvents) {
-            const start = moment(calendarEvent.dates.startDate);
-
-            if (start.isSame(calendarEvent.dates.endDate, 'day')) {
-                intervalsModel.set(start, {
-                    intervalType: 'intervalBoundary',
-                    eventType: calendarEvent.type
-                });
-                continue;
-            }
-
-            while (start.isSameOrBefore(calendarEvent.dates.endDate, 'day')) {
-                let intervalType: IntervalType = 'interval';
-
-                if (start.isSame(calendarEvent.dates.startDate)) {
-                    intervalType = 'startInterval';
-                } else if (start.isSame(calendarEvent.dates.endDate)) {
-                    intervalType = 'endInterval';
-                }
-
-                intervalsModel.set(start, {
-                    intervalType: intervalType,
-                    eventType: calendarEvent.type
-                });
-
-                start.add(1, 'days');
-            }
-        }
-
-        return intervalsModel;
     }
 }
